% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_annot}
\alias{add_annot}
\title{add_annot}
\usage{
add_annot(qpcr, file)
}
\arguments{
\item{qpcr}{An instance of the qpcr class.}

\item{file}{A path to a sample annotation file. This file should at least have one column named 'Sample'.}
}
\value{
An qpcr object with added sample annotation.
}
\description{
Add sample annotation
}
\details{
\code{add_annot} allows one to add a sample annotation file to the qpcr object.
}
