% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{analyse_qpcr}
\alias{analyse_qpcr}
\title{analyse_qpcr}
\usage{
analyse_qpcr(
  qpcr,
  reference_genes,
  levels = NULL,
  aggregation = c("geomean", "mean", "median")
)
}
\arguments{
\item{qpcr}{An instance of the qpcr class. Expression data and sample annotations must have been loaded already.}

\item{reference_genes}{A vector listing the genes one wishes to use as housekeeping genes or reference genes.}

\item{levels}{A list with the set levels for each factor in the sample annotation data. Defaults to NULL.}

\item{aggregation}{Defines how CT values of technical replicates and reference genes should be aggregated; either by their arithmetic ('mean') or geometric mean ('geomean') or their median ('median').}
}
\value{
An updated instance of the qpcr class with analysed expression data.
}
\description{
Analyse qpcr data
}
\details{
\code{analyse_qpcr} allows one to perform the core analysis of qpcr data.
}
