% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CombineTechReps.R
\docType{methods}
\name{CombineTechReps}
\alias{CombineTechReps}
\title{Combine technical replicates of Cq and amplification efficiencies}
\usage{
\S4method{CombineTechReps}{RTqPCRBatch}(RTqPCRBatch, calc = "Mean",
  cRepCq = TRUE, ...)
}
\arguments{
\item{RTqPCRBatch}{Object of class \code{RTqPCRBatch}. It can be output of \code{\link[RTqPCR]{CqValues}} or
\code{\link[RTqPCR]{ReplaceValue}} or \code{\link[RTqPCR]{NonDetects}} or \code{\link[RTqPCR]{ReplaceAboveCutOff}}
or \code{\link[RTqPCR]{ReplaceNAs}} functions.}

\item{calc}{Which methods to be used out of Mean, Median and Geometric mean.}

\item{cRepCq}{if TRUE, then combine replicates in Cq. If FALSE, then combine replicates in amplification 
efficiencies.}

\item{\dots}{Other parameters to be passed to downstream methods.}
}
\value{
Object of class \code{"RTqPCRBatch"} with new slots.
}
\description{
Combine the technical replicates of Cq values and amplification efficiencies on the basis 
of mean, median and geometric mean and compute standard deviation of Cq and efficiency within the replicates
in object of class \code{"RTqPCRBatch"}.
}
\details{
See the vignettes in \pkg{RTqPCR} package.
}
\examples{

## Read in the raw fluorescent data

LC480.example <- file.path(path, "LC480_Example.txt") 
cycData.LC480 <- read.RTqPCR(LC480.example, PCRtype = "LC480")

## Read in the sample information data

SampleInfoLC480 <- file.path(path, "LC480_example_SampleInfo.txt")
samInfoLC480 <- read.RTqPCRSampleInfo(SampleInfoLC480, PCRtype = "LC480")

## Merge the fluorescence and sample information data through merge function

merge.LC480<-merge(cycData.LC480,samInfo.LC480) 

## Compute the Cq values and amplification efficiencies

Cqeffs.LC480 <- CqValues(merge.LC480, PCRtype = "LC480", Effmethod = "sigfit", baseline = "none")
Cqeffs.LC480 #To see the overview of data
exprs(Cqeffs.LC480)[1:5]  ##to visualise the first five CqValues
effs(Cqeffs.LC480)[1:5]  ##to visualise the first five amplification efficiencies
exprs(Cqeffs.LC480)  ##to visualise all Cq values
effs(Cqeffs.LC480)   ##to visualise all amplification efficiencies

##  Implement the functions ReplaceNAs, ReplaceAboveCutOff, ReplaceValue and NonDetects as per requirement.

## Combine technical replicates - In this example, we are combining the replicates directly from 
## the output of CqValues function

##  replicates can be combined through following three methods

## 1) based on mean (default method is for Cq values and is based on mean)

## i) combine Cq values and compute the standard deviation of Cq values within replicates

Cqreps <- CombineTechReps(Cqeffs.LC480)
Cqreps   ## to visualise the overview of the rsulting data


ii) combine amplification efficiencies and compute the standard deviation of amplification efficiencies
within replicates

effsreps <- CombineTechReps(Cqeffs.LC480, cRepCq = FALSE)

## 2) Based on median

## i) Combine Cq values and compute standard deviation of Cq values within replicates

Cqreps2 <- CombineTechReps(Cqeffs.LC480, calc = "Median")

ii) Combine amplification efficiencies and compute their standard deviation within replicates

effsreps2 <- CombineTechReps(Cqeffs.LC480, calc = "Median", cRepCq = FALSE)

## 3) Based on Geometric Mean

## i) combine Cq values and compute their standard deviation within replicates

Cqreps3 <- CombineTechReps(Cqeffs.LC480, calc = "Geom")

ii) combine amplification efficiencies and compute their standard deviation within replicates

effsreps <- CombineTehcReps(Cqeffs.LC480, calc = "Geom", cRepCq = FALSE)

}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
