% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaCq.R
\docType{methods}
\name{DeltaCq}
\alias{DeltaCq}
\title{Compute delta Cq value and standard deviation of delta Cq}
\usage{
\S4method{DeltaCq}{RTqPCRBatch}(RTqPCRBatch, Ref, ...)
}
\arguments{
\item{RTqPCRBatch}{An object of class \code{"RTqPCRBatch"} produced as an output of CombineTechReps function.}

\item{Ref}{Name of reference gene under column Target name.}

\item{\dots}{other parameters to be passed to the downstream methods.}
}
\value{
An object of \code{"RTqPCRBatch"} with new slots.
}
\description{
Compute delta Cq and standard deviation of delta Cq of object of \code{"RTqPCRBatch"}.
}
\examples{

## Read in the raw fluorescent data

## Read in the sample information data

## Merge the fluorescence and sample information data through merge function

## Compute the Cq values and amplification efficiencies

##  Implement the functions ReplaceNAs, ReplaceAboveCutOff, ReplaceValue and NonDetects as per requirement.

## Combine technical replicates (based on Cq values) : see also CombineTechReps function and vignettes

Cqreps <- CombineTechReps(Cqeffs.LC480)
Cqreps   ## to visualise the overview of the resulting data

## To compute the delta Cq value
deltaCq <- DeltaCq(Cqreps, Ref = "hk")

## For more details see the vignettes

}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
