% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaDeltaCq.R
\docType{methods}
\name{DeltaDeltaCq}
\alias{DeltaDeltaCq}
\title{Compute delta delta Cq, standard deviation of delta delta Cq and fold concentration.}
\usage{
\S4method{DeltaDeltaCq}{RTqPCRBatch}(RTqPCRBatch, Target, ...)
}
\arguments{
\item{RTqPCRBatch}{An object of class \code{"RTqPCRBatch"} produced as an output of DeltaCq function.}

\item{Target}{Name of target gene under column Target name.}

\item{\dots}{other parameters to be passed to the downstream methods.}
}
\value{
An object of \code{"data.frame"} with new slots.
}
\description{
Compute delta delta Cq, standard deviation of delta delta Cq and fold concentration of an 
object of \code{"RTqPCRBatch"}.
}
\examples{

## Read in the raw fluorescent data

## Read in the sample information data

## Merge the fluorescence and sample information data through merge function

## Compute the Cq values and amplification efficiencies

##  Implement the functions ReplaceNAs, ReplaceAboveCutOff, ReplaceValue and NonDetects as per requirement.

## Combine technical replicates (based on Cq values) : see also CombineTechReps function and vignettes

Cqreps <- CombineTechReps(Cqeffs.LC480)
Cqreps   ## to visualise the overview of the resulting data

## To compute the delta Cq value
## To compute the delta delta Cq value implement the DeltaDeltaCqAll function. It's an auxilliary function
##to DeltaDeltaCqAll function
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
