% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeltaDeltaCqAll.R
\docType{methods}
\name{DeltaDeltaCqAll}
\alias{DeltaDeltaCqAll}
\title{Compute delta delta Cq, standard deviation of delta delta Cq and fold concentration of all genes 
at one time.}
\usage{
\S4method{DeltaDeltaCqAll}{RTqPCRBatch}(RTqPCRBatch, ...)
}
\arguments{
\item{\dots}{other parameters to be passed to the downstream methods.}

\item{x}{An object of class \code{"RTqPCRBatch"} produced as an output of DeltaCq function.}
}
\value{
An object of \code{"RTqPCRBatch"} with new slots.
}
\description{
Compute delta delta Cq, standard deviation of delta delta Cq and fold concentration of an 
object of \code{"RTqPCRBatch"}.
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
#'@examples 

## Read in the raw fluorescent data

## Read in the sample information data

## Merge the fluorescence and sample information data through merge function

## Compute the Cq values and amplification efficiencies

##  Implement the functions ReplaceNAs, ReplaceAboveCutOff, ReplaceValue and NonDetects as per requirement.

## Combine technical replicates (based on Cq values) : see also CombineTechReps function and vignettes

Cqreps <- CombineTechReps(Cqeffs.LC480)
Cqreps   ## to visualise the overview of the resulting data

## To compute the delta Cq value
deltaCq <- DeltaCq(Cqreps, Ref = "hk")
deltaCq

## For more details see the vignettes
deltadeltaCqall <- DeltaDeltaCqAll(deltaCq)
deltadeltaCqall
}
