% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRQeffs.R
\docType{methods}
\name{NRQeffs}
\alias{NRQeffs}
\title{Compute relative expression ratio}
\usage{
\S4method{NRQeffs}{RTqPCRBatch}(RTqPCRBatch, Target, Ref, ...)
}
\arguments{
\item{Target}{Name of target sample gene under column Target name.}

\item{Ref}{Name of reference gene under column Target name.}

\item{\dots}{Other parameters to be passed to the downstream methods.}

\item{x}{An object of class \code{"RTqPCRBatch"} produced as an result of function CombineTechReps.}
}
\value{
An object of class \code{"data.frame"} with new slots.
}
\description{
Perform normalization and compute relative expression ratio of target sample on the 
basis of efficiency based methods of Roche, Pfaffl and delta delta Cq from an object of class \code{"RTqPCRBatch"}.
}
\examples{

## Read in the raw fluorescent data

## Read in the sample information data

## Merge the fluorescence and sample information data through merge function

## Compute the Cq values and amplification efficiencies

##  Implement the functions ReplaceNAs, ReplaceAboveCutOff, ReplaceValue and NonDetects as per requirement.

## Combine technical replicates - on the basis of efficiency

effsreps <- CombineTechReps(Cqeffs.LC480, cRepCq = FALSE)

##NRQeffs is an auxiliarry function to NRQeffsAll function. It can be implemented directly through the 
##NRQeffsAll function

}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
