% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NRQeffsAll.R
\docType{methods}
\name{NRQeffsAll}
\alias{NRQeffsAll}
\title{Compute normalized relative expression of all genes at one time.}
\usage{
\S4method{NRQeffsAll}{RTqPCRBatch}(RTqPCRBatch, y, ...)
}
\arguments{
\item{RTqPCRBatch}{An object of class \code{"RTqPCRBatch"} produced as an output of CombineTechReps function.}

\item{y}{Name of Reference gene under column Target name.}

\item{\dots}{other parameters to be passed to the downstream methods.}
}
\value{
An object of \code{"RTqPCRBatch"} with new slots.
}
\description{
Compute relative expression ration of all genes at one time of an 
object of \code{"RTqPCRBatch"}.
}
\examples{

## Read in the raw fluorescent data

## Read in the sample information data

## Merge the fluorescence and sample information data through merge function

## Compute the Cq values and amplification efficiencies

##  Implement the functions ReplaceNAs, ReplaceAboveCutOff, ReplaceValue and NonDetects as per requirement.

## Combine technical replicates - on the basis of efficiency

effsreps <- CombineTechReps(Cqeffs.LC480, cRepCq = FALSE)

##To compute the relative expression ratio
nrqeffsall <- NRQeffsAll(effsreps, y = "hk")
nrqeffsall

}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
