% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{qcpr-class}
\alias{qcpr-class}
\alias{qpcr}
\title{An S4 class to contain and process RT-qPCR data}
\description{
An S4 class to contain and process RT-qPCR data
}
\section{Slots}{

\describe{
\item{\code{ct}}{A dataframe containing the raw CT values.}

\item{\code{references}}{A dataframe containing the reference expression data.}

\item{\code{dct}}{A dataframe containing the delta CT values}

\item{\code{annotation}}{A dataframe containing the sample annotation data.}

\item{\code{melting}}{A dataframe containing the melting curve data of each sample.}
}}

