% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMx3005P.R
\name{read.Mx3005P}
\alias{read.Mx3005P}
\title{To read in the raw fluorescent data of Mx3005P (Agilent) RT-qPCR experiments}
\usage{
read.Mx3005P(file, colNames = c("mxpSegment", "Ramp/Plateaue",
  "Ramp/Plateaue#", "Well", "Dye", "Cycle#", "Fluorescence", "Temperature"),
  cycleThreshold = 45, fileType = "txt", skip = 1, header = TRUE,
  sep = "\\t", quote = "\\"", dec = ".", fill = TRUE,
  comment.char = "")
}
\arguments{
\item{file}{the name of the file to read in.}

\item{colNames}{a character vector of names to be assumed for the columns.}

\item{cycleThreshold}{maximum number of cycles which will be read in.}

\item{fileType}{the type of the file.}

\item{skip}{an integer: the number of lines of the data file to skip before beginning to read data.}

\item{header}{a logical value indicating whether the file contains the names of the variables 
as its first line. If missing, the value is determined from the file format: header is set to TRUE
if and only if the first row contains one fewer field than the number of columns.}

\item{sep}{the field separator character. Values on each line of the file are separated by this character.
See \code{\link[utils]{read.table}}.}

\item{quote}{the set of quoting characters. To disable quoting altogether, use \code{quote = ""}.
See \code{\link[base]{scan}} for the behaviour on quotes embedded in quotes. Quoting is only considered 
for columns read as character, which is all of them unless colClasses is specified.}

\item{dec}{the character used in the file for decimal points.}

\item{fill}{logical. If TRUE then in case the rows have unequal length, blank fields are implicitly added.
See \code{\link[utils]{read.table}}.}

\item{comment.char}{character: a character vector of length one containing a single character or an empty string. 
Use \code{""} to turn off the interpretation of comments altogether.}

\item{\dots}{other parameters to be passed to downstream methods.}
}
\value{
object of Class \code{"RTqPCRBatch"} with exprs slots.
}
\description{
Function \code{read.Mx3005P} reads in txt files of raw fluorescent data of Mx3005P 
RT -qPCR and uses the data to populate an object of class \code{"RTqPCRBatch"}.
}
\details{
Allows the user to read in qPCR raw fluorescence data from Mx3005P RT-qPCR which has been exported 
to a txt-file, alongside phenotypic data. More details on how to read in raw fluorescence data from RT-qPCR
or Light Cycler in the \code{RTqPCR} package vignette.
}
\examples{
path <- system.file("exData", package = "RTqPCR")
Mx3005P.example <- file.path(path, "Mx3005P_Example.txt")
cycData <- read.Mx3005P(Mx3005P.example)
cycData[1:5] ##to visualise the first 5 cycdata and for all data remove [1:5] from the code
featureData(cycData)  ## to visualise the overview of the featureData (feature information of data) 
fData(cycData)[1:5] ## To visualise the content of the first five featureData
## fData(cycData)   To visualise content of all featureData
phenoData(cycData) ## To visualise the overview of phenoData (phenotypic information of data)
pData(cycData)[1:5] ## To visualise the content of the first five phenoData
## pData(cycData) To visualise the content of all phenoData
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{CyclesSet}
