% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRTqPCR.R
\name{read.RTqPCR}
\alias{read.RTqPCR}
\title{To read in the raw fluorescent data of LC480 light cycler and Mx3005P RT-qPCR experiment.}
\usage{
read.RTqPCR(file, PCRtype, ...)
}
\arguments{
\item{file}{the name of the file to read in.}

\item{PCRtype}{the type of RT-qPCRs ("LC480" or "Mx3005P") whose file is to read in.}

\item{\dots}{Other parameters to be passed to downstream methods.}
}
\value{
\code{"RTqPCRBatch"} object with exprs slots.
}
\description{
Function \code{read.RTqPCR} reads in the raw fluorescent data of LC480 light cycler and Mx3005P
RT-qPCR and use the data to populate an object of class \code{"RTqPCRBatch"}.
}
\details{
Function \code{read.RTqPCR} reads in the raw fluorescent data of LC480 light cycler
and Mx3005P RTqPCR and is based on functions \code{\link[ReadqPCR]{read.LC480}} and \code{\link{read.Mx3005P}}.
}
\examples{
##To read in the fluorescent data of LC480 light cycler

LC480.example <- file.path(path, "LC480_Example.txt") 
rtData.LC480 <- read.RTqPCR(LC480.example, PCRtype = "LC480")
rtData.LC480 ## to visualise the overview of data
head(exprs(rtData.LC480)) ## to visualise all of the fluorescence data
head(exprs(rtData.LC480[,1:5])) ## to visualise the first five columns of fluorescence data
head(pData(rtData.LC480)) ## to express the phenoData 
head(fData(rtdata.LC480)) ## to express featureData

##To read in the fluorescent data of Mx3005P RT-qPCR

Mx3005P.example <- file.path(path, "Mx3005P_Example.txt") 
rtData.Mx <- read.RTqPCR(Mx3005P.example, PCRtype = "Mx3005P")
rtData.Mx ## to visualise the overview of data
head(exprs(rtData.Mx)) ## to visualise all of the fluorescence data
head(exprs(rtData.Mx[,1:5])) ## to express the first five columns of fluorescence data
head(pData(rtData.Mx)) ## to express the phenoData
head(fData(rtData.Mx)) ## to express the featureData

}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
