% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_qpcr}
\alias{read_qpcr}
\title{read_qpcr}
\usage{
read_qpcr(qpcr, files, keep_undetected = TRUE)
}
\arguments{
\item{qpcr}{An empty instance of the qpcr class.}

\item{files}{A vector of paths pointing to qPCR raw data files.}

\item{keep_undetected}{A logical indicating whether to keep undetected CT values and set them to 40, or to simply delete those entries. Defaults to TRUE.}
}
\value{
An updated instance of the qpcr class.
}
\description{
read_qpcr
}
