\name{RTqPCR-package}
\alias{RTqPCR-package}
\alias{RTqPCR}
\docType{package}
\title{
Work Flow of RT-qPCR
}
\description{
The work flow provides functions and graphical user interface to compute the complete work
flow of LC480 and Mx3005P RT-qPCR
}
\details{
\tabular{ll}{
Package: \tab RTqPCR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-04-09\cr
License: \tab LGPL-3\cr
Depends: \tab R(>= 3.0.1), Biobase, methods, affy, qpcR, stats, RColorBrewer\cr
Imports: \tab Biobase\cr
Suggests: \tab qpcR\cr
Lazyload: \tab Yes\cr
}
library(RTqPCR)
}
\author{
Navneet Phogat, Matthias Kohl <Matthias.Kohl@stamats.de>\cr

Maintainer: Matthias Kohl <Matthias.Kohl@stamats.de>\cr
}
\references{
~~ Literature or other references for background information ~~
}

\examples{
Some examples are described in vignette.
\dontrun{
library(LC480)
vignette("LC480")
}
}
\keyword{ package }
