% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTqPCR.dataframe.R
\docType{methods}
\name{RTqPCR.dataframe}
\alias{RTqPCR.dataframe}
\title{Convert the object of Class \code{"data.frame"} into an object of Class \code{"RTqPCRBatch"}.}
\usage{
\S4method{RTqPCR.dataframe}{data.frame}(x, fun, pcrtype, ...)
}
\arguments{
\item{x}{An object of Class \code{"data.frame"}.}

\item{fun}{Name of function whose output is to be converted into RTqPCRBatch.}

\item{\dots}{other parameters to be passed to the downstream methods.}
}
\value{
An object of Class \code{"RTqPCRBatch"} with new slots.
}
\description{
Convert the object of Class \code{"data.frame"} into an object of Class \code{"RTqPCRBatch"}.
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
\keyword{RTqPCRBatch}
\keyword{data.frame}
