% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTqPCR.gui.R
\name{RTqPCR.gui}
\alias{RTqPCR.gui}
\title{Run the graphical User Interface}
\usage{
RTqPCR.gui()
}
\description{
Run the graphical user interface
}
\examples{
RTqPCR.gui()
}
\author{
Navneet Phogat, Matthias Kohl, \email{Matthias.Kohl@stamats.de}
}
